Rename `unwind.rs` to reflect its contents.
Rename it to `unwind_unimplemented.rs` so that the code in lib.rs is
more self-describing.
